Make lang file paths relative to scripts, add checks for motd, fix formatting

+++ setup.sh
@@ -13,7 +13,7 @@ set -e
 # set -x  # Uncomment for debugging
 
 # Variables
-WORKING_DIR="${HOME}/.termux"
+WORKING_DIR="$(dirname "${BASH_SOURCE[0]}")"
 LANG_CODE="en"
 
 # Pre-scan all arguments for -l <lang_code> before loading lang files
@@ -29,7 +29,7 @@ for ((i = 1; i <= "${#}"; i++)); do
 			exit 1
 		fi
 		lang_input="${next_arg,,}"
-		if [[ -d "${WORKING_DIR}/lang/${lang_input}" ]]; then
+		if [[ -d "${WORKING_DIR}/Termux/lang/${lang_input}" ]]; then
 			LANG_CODE="${lang_input}"
 		else
 			printf "${red}No localization found for language code${nocol} ${lang_input}\n"
@@ -41,14 +41,14 @@ for ((i = 1; i <= "${#}"; i++)); do
 done
 
 # Load lang files
-mapfile -t LANG_STRINGS < "${WORKING_DIR}/lang/${LANG_CODE}/setup.lang"
-mapfile -t COMMON_STRINGS < "${WORKING_DIR}/lang/${LANG_CODE}/common.lang"
+mapfile -t LANG_STRINGS < "${WORKING_DIR}/Termux/lang/${LANG_CODE}/setup.lang"
+mapfile -t COMMON_STRINGS < "${WORKING_DIR}/Termux/lang/${LANG_CODE}/common.lang"
 
 while [[ ${#} -gt 0 ]]; do
 	case "${1}" in
 		-ls)
 			printf "${green}${COMMON_STRINGS[9]}${nocol}:\n"
-			find "${WORKING_DIR}/lang" -mindepth 1 -maxdepth 1 -type d -exec basename {} \;
+			find "${WORKING_DIR}/Termux/lang" -mindepth 1 -maxdepth 1 -type d -exec basename {} \;
 			exit 0
 			;;
 		-h)
@@ -66,60 +66,64 @@ while [[ ${#} -gt 0 ]]; do
 done
 
 install_dependencies() {
-	printf "${green}${LANG_STRINGS[0]}...${nocol}"
+	printf "${green}${LANG_STRINGS[0]}...${nocol}\n"
 	apt update && apt install -y fontconfig-utils git zsh figlet toilet lf curl wget micro man || {
-		printf "${red}${LANG_STRINGS[1]}!${nocol}"
+		printf "${red}${LANG_STRINGS[1]}!${nocol}\n"
 		exit 1
 	}
 }
 
 configure_termux() {
-	printf "${green}${LANG_STRINGS[2]}...${nocol}"
+	printf "${green}${LANG_STRINGS[2]}...${nocol}\n"
 	if [ -d "${HOME}/.termux" ]; then
-		printf "${green}${LANG_STRINGS[3]}${nocol}: ${HOME}/.termux_bak"
+		printf "${green}${LANG_STRINGS[3]}${nocol}: ${HOME}/.termux_bak\n"
 		mv "${HOME}/.termux" "${HOME}/.termux_bak"
 	fi
 	cp -r Termux "${HOME}/.termux" || {
-		printf "${red}${LANG_STRINGS[4]}!${nocol}"
+		printf "${red}${LANG_STRINGS[4]}!${nocol}\n"
 		exit 1
 	}
 	chmod +x "${HOME}/.termux/fonts.sh" "${HOME}/.termux/colors.sh"
-	printf "${green}${LANG_STRINGS[5]}...${nocol}"
+	printf "${green}${LANG_STRINGS[5]}...${nocol}\n"
 	ln -fs "${HOME}/.termux/colors/dark/IrBlack" "${HOME}/.termux/colors.properties"
 	# Replacing termuxs boring welcome message with something good looking
-	mv "${PREFIX}/etc/motd" "${PREFIX}/etc/motd.bak"
-	mv "${PREFIX}/etc/motd.sh" "${PREFIX}/etc/motd.sh.bak"
+	if [[ -f "${PREFIX}/etc/motd" ]]; then
+	    mv "${PREFIX}/etc/motd" "${PREFIX}/etc/motd.bak"
+	fi
+	if [[ -f "${PREFIX}/etc/motd.sh" ]]; then
+	    mv "${PREFIX}/etc/motd.sh" "${PREFIX}/etc/motd.sh.bak"
+	fi
 	ln -sf "${HOME}/.termux/motd.sh" "${PREFIX}/etc/motd.sh"
 }
 
 install_ohmyzsh() {
-	printf "${green}${LANG_STRINGS[6]}...${nocol}"
+	printf "${green}${LANG_STRINGS[6]}...${nocol}\n"
 	git clone https://github.com/ohmyzsh/ohmyzsh.git "${HOME}/.oh-my-zsh"
-	printf "${green}${LANG_STRINGS[7]}...${nocol}"
+	printf "${green}${LANG_STRINGS[7]}...${nocol}\n"
 	git clone --depth=1 https://github.com/romkatv/powerlevel10k.git "${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/themes/powerlevel10k"
-	printf "${green}${LANG_STRINGS[8]}...${nocol}"
+	printf "${green}${LANG_STRINGS[8]}...${nocol}\n"
 	git clone https://github.com/zsh-users/zsh-autosuggestions.git "${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/plugins/zsh-autosuggestions"
 	git clone https://github.com/zsh-users/zsh-syntax-highlighting.git "${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting"
-	printf "${green}${LANG_STRINGS[9]}...${nocol}"
+	printf "${green}${LANG_STRINGS[9]}...${nocol}\n"
 	cp -f OhMyZsh/zshrc "${HOME}/.zshrc"
 	if [[ "$(dpkg --print-architecture)" == "arm" ]]; then
-		printf "${red}${LANG_STRINGS[10]}!${nocol}"
+		printf "${green}${LANG_STRINGS[10]}!${nocol}\n"
 		# There's no binaries of gitstatus for armv7 right now so disable it
 		printf "\n# Disable gitstatus for now (Only for armv7 devices)\nPOWERLEVEL9K_DISABLE_GITSTATUS=true\n" >> "${HOME}/.zshrc"
 	fi
 	chmod +rwx "${HOME}/.zshrc"
 	if [[ -f "OhMyZsh/zsh_history" ]]; then
-		printf "${green}${LANG_STRINGS[11]}...${nocol}"
+		printf "${green}${LANG_STRINGS[11]}...${nocol}\n"
 		cp -f OhMyZsh/zsh_history "${HOME}/.zsh_history"
 		chmod +rw "${HOME}/.zsh_history"
 	fi
 	if [[ -f "OhMyZsh/custom_aliases.zsh" ]]; then
-		printf "${green}${LANG_STRINGS[12]}...${nocol}"
+		printf "${green}${LANG_STRINGS[12]}...${nocol}\n"
 		cp -f OhMyZsh/custom_aliases.zsh "${ZSH_CUSTOM:-${HOME}/.oh-my-zsh/custom}/custom_aliases.zsh"
 	fi
-	printf "${green}${LANG_STRINGS[13]}...${nocol}"
+	printf "${green}${LANG_STRINGS[13]}...${nocol}\n"
 	cp -f OhMyZsh/p10k.zsh "${HOME}/.p10k.zsh"
-	printf "${green}${LANG_STRINGS[14]}!${nocol}"
+	printf "${green}${LANG_STRINGS[14]}!${nocol}\n"
 }
 
 finish_install() {
@@ -133,18 +137,18 @@ finish_install() {
 	if [[ "$(dpkg --print-architecture)" == "arm" ]]; then
 		rm -rf "${HOME}/.cache/gitstatus"
 	fi
-	printf "${green}${LANG_STRINGS[15]}...${nocol}"
+	printf "${green}${LANG_STRINGS[15]}...${nocol}\n"
 	chsh -s zsh
 	# Setup Complete
 	termux-setup-storage
 	termux-reload-settings
-	printf "${green}${LANG_STRINGS[16]}!${nocol}"
-	printf "${green}${LANG_STRINGS[17]}!${nocol}"
+	printf "${green}${LANG_STRINGS[16]}!${nocol}\n"
+	printf "${green}${LANG_STRINGS[17]}!${nocol}\n"
 }
 
 main() {
 	# Start installation
-	printf "${green}${LANG_STRINGS[18]} Termux-Zsh${nocol}?\n"
+	printf "${green}${LANG_STRINGS[18]} Termux-zsh?${nocol}\n"
 	printf "  [${green}1${nocol}] ${COMMON_STRINGS[3]}\n"
 	printf "  [${green}2${nocol}] ${COMMON_STRINGS[6]}\n"
